fix(main): resolved compiler warning, type mismatches and improved result handling - Promoted float to double, updated scanf specifiers to %lf and function signature - Initialized result to NAN, reset after each iteration - Gated displayhistory() with !isnan(result)

+++ include/history.h
@@ -4,6 +4,6 @@
 // Defining maximum history size
 #define MAX_HISTORY 5 
 
-void displayHistory(double result, float history[], int *hCount); // Displays the recent history of results
+void displayHistory(double result, double history[], int *hCount); // Displays the recent history of results
 
 #endif
